% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{addMyShellPath}
\alias{addMyShellPath}
\title{addMyShellPath}
\usage{
addMyShellPath(path = NULL)
}
\description{
Adds extra folder to R path if necessary. Used to implement python, which can
convert Excel files to .csv files (which R can read). If this function does not work,
the best solution is to open the Excel file and save it as a .csv file. Then tell
R to load the .csv file instead.
}

