% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{allEqual}
\alias{allEqual}
\title{All Equal}
\usage{
allEqual(v, error = FALSE)
}
\arguments{
\item{v}{vector}

\item{error}{if True, stops the program execution}
}
\value{
value of each item in the vector if True, else NA
}
\description{
Checks if all values in a vector are the same. Uses assert function.
}

