% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisFunctions.R
\name{bydf}
\alias{bydf}
\alias{byvector}
\title{By: For Data Frames and Vectors}
\usage{
bydf(...)

byvector(...)
}
\description{
Wrapper functions for by().
}
\details{
bydf converts the output to a data frame. It should only be used if the
output of FUN is a length-1 vector. Creates a data frame with 1 column per
index, plus a 'Freq' column that is the output of FUN.

byvector converts the output to a vector. This is likely useful when calling
`by` with only 1 index and when the output of FUN is always the same length.
}

