% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_subfunctions.R
\name{calculateCommGlasso}
\alias{calculateCommGlasso}
\title{Calculating the degree of each community in a network}
\usage{
calculateCommGlasso(
  S,
  x,
  netlist,
  lambda.max = 0.9,
  lambda = NULL,
  scale = F,
  additional = NULL,
  verbose = T
)
}
\arguments{
\item{net}{LR network}

\item{communites}{Communities of interest}

\item{dat}{Dataset of one cell type}
}
\value{
List of maximum degree in each community
}
\description{
Used to assess significance of the network relative to the
sample size of the dataset
}
