% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{calculateCor}
\alias{calculateCor}
\title{calculateCor - calculating pairwise correlation for all the LR pairs}
\usage{
calculateCor(lr.table, dat.list)
}
\arguments{
\item{lr.table}{List of cell type-specific LR pairs}

\item{dat.list}{List of gene expression for each cell type}
}
\value{
A table of correlations for each LR pair
}
\description{
This is an internal function used in generateLRnet. It calculates
Pearson correlation of all ligand receptor pairs.
}
