% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_subfunctions.R, R/load_subfunctions_old.R
\name{calculateNetVars}
\alias{calculateNetVars}
\title{Calculating the degree of each community in a network}
\usage{
calculateNetVars(net, genes)

calculateNetVars(net, genes)
}
\arguments{
\item{net}{LR network}

\item{communites}{Communities of interest}

\item{dat}{Dataset of one cell type}
}
\value{
List of maximum degree in each community

List of maximum degree in each community
}
\description{
Used to assess significance of the network relative to the
sample size of the dataset

Used to assess significance of the network relative to the
sample size of the dataset
}
