% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_subfunctions.R
\name{calculateOversizedComms}
\alias{calculateOversizedComms}
\title{Calculating the degree of each community in a network}
\usage{
calculateOversizedComms(net, communities, dat, maxNum)
}
\arguments{
\item{net}{LR network}

\item{dat}{Dataset of one cell type}

\item{communites}{Communities of interest}
}
\value{
List of maximum degree in each community
}
\description{
Used to assess significance of the network relative to the
sample size of the dataset
}
