% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_subfunctions.R
\name{calculatePvalue}
\alias{calculatePvalue}
\title{Calculates p-value for a community}
\usage{
calculatePvalue(R_, S_, D_, i_, j_, n, p)
}
\arguments{
\item{R_}{correlation matrix}

\item{S_}{covariance matrix}

\item{D_}{list of correlation perturbations and number of times graphical lasso predicted an edge to be present or not}

\item{i_}{index for ligand}

\item{j_}{index for receptor}

\item{n}{sample size}

\item{p}{number of features in community}
}
\value{
p-value
}
\description{
Calculates p-value for a community
}
