% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{calculateSignificance}
\alias{calculateSignificance}
\title{Calculating p-value for a given edge in a community}
\usage{
calculateSignificance(
  obj,
  l.chosen,
  r.chosen,
  comm.chosen,
  maxNum,
  iterNum = 1000,
  seednum = 30,
  lambda = NULL
)
}
\arguments{
\item{obj}{REMI object}

\item{l.chosen}{Ligand in edge}

\item{r.chosen}{Receptor in edge}

\item{comm.chosen}{Community that edge is present in}

\item{iterNum}{= Number of permutations for p-value calculation}

\item{seednum}{= Seed}

\item{lambda}{= Manually setting lambda if of interest}
}
\value{
Chord Diagram highlighting proportion of cell-types in interactome
}
\description{
Calculating p-value for a given edge in a community
}
