% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{ChordPlot}
\alias{ChordPlot}
\title{Making chord diagram for caberNET results}
\usage{
ChordPlot(interactome, grid.col)
}
\arguments{
\item{interactome}{Predicted interactome}

\item{grid.col}{Optional: color option for cell type}
}
\value{
Chord Diagram highlighting proportion of cell-types in interactome
}
\description{
Making chord diagram for caberNET results
}
