% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_subfunctions.R
\name{clusterLouvain}
\alias{clusterLouvain}
\title{Clustering using Louvain}
\usage{
clusterLouvain(net, commnums, communities)
}
\arguments{
\item{net}{LR network}

\item{commnums}{List of communities that had a degree larger than sample size}

\item{communities}{List of all communities in LR network}
}
\value{
Clusters identified using louvain combined with all communities
}
\description{
The large communities identified by label propagation are broken down
into even smaller communities to ensure the degree of the networks
reflect the size of the dataset to maximize significance.
}
