% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysisFunctions.R
\name{corstarsl}
\alias{corstarsl}
\title{corstarsl}
\usage{
corstarsl(x)
}
\arguments{
\item{x}{matrix-like R object}
}
\value{
data frame of correlations with significance ratings
}
\description{
Creates a correlation table with stars for significance!!!
}
\details{
*** = 0.1%   ** = 1%   * = 5%
}

