% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{createAnalysisFolders}
\alias{createAnalysisFolders}
\title{Make Default Analysis Folders}
\usage{
createAnalysisFolders(mainFolder, addCSS = FALSE)
}
\arguments{
\item{mainFolder}{main analysis folder}
}
\value{
data frame of subfolders
}
\description{
Creates EMI analysis folder template, including subfolders for:
}
\details{
- Input Data
- Figures
- Working
}

