% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisFunctions.R
\name{duplicateInfo}
\alias{duplicateInfo}
\title{Duplicate Information}
\usage{
duplicateInfo(x, missingVals = c("", NA))
}
\arguments{
\item{x}{vector of interest}

\item{missingVals}{values within x to consider as missing. Defaults to
considering empty strings and NAs as missing}
}
\description{
Determines the number of unique, duplicate, and missing values in a vector
}

