% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingFunctions.R
\name{EMIbar}
\alias{EMIbar}
\title{EMI Bar Graph 1-D}
\usage{
EMIbar(data, colors = EMIcolors(1), dataCols = 1:2, ...)
}
\arguments{
\item{data}{a data frame with columns X (bar names), and Count (bar size).
Works best with data from the myTable function.}

\item{colors}{vector of colors with length of 1 or the same length as data.}

\item{dataCols}{length 2 vector specifying the location of data within the
data frame. The first column should have category info, the second should have
counts.
Default colors are all EMI Blue (EMIcolors(1))}
}
\description{
Creates generic EMI bar graph. Takes a data frame (or data from myTable) to
create a 1-D plot. Colors are customizable.
}
\examples{
df = data.frame(Factor1 = sample(LETTERS[1:4], 100, replace = T),
                Factor2 = sample(LETTERS[25:26], 100, replace = T),
                Numbers = sample(1:100))

# EMIbar
t = myTable(as.table(by(df$Numbers, df$Factor1, mean)))
names(t)
EMIbar(t)
}

