% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingFunctions.R
\name{EMIbarMulti}
\alias{EMIbarMulti}
\title{EMI Bar Graph 2-D}
\usage{
EMIbarMulti(data, colors = NULL)
}
\arguments{
\item{data}{a data frame with columns X (bar names), Fill (bar color),
and Count (bar size). Works best with data from the myTableMulti function.}

\item{colors}{vector of colors with length of 1 or the same length as the
number of Fill options. Default colors are all EMI Blue (EMIcolors(1))}

\item{...}{parameters passed to geom_bar}
}
\description{
Creates generic EMI bar graph. Takes a data frame (or data from myTableMulti) to
create a 2-D plot. Colors are customizable.
}

