% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingFunctions.R
\name{EMIcolors}
\alias{CEcolors}
\alias{EMIcolors}
\alias{Navigantcolors}
\alias{Navigantcolors2}
\title{EMI Color Scheme}
\usage{
EMIcolors(colors = 1, accentLoc = NA, ramp = 0, accent = NULL,
  order = NULL)

Navigantcolors(...)

Navigantcolors2(...)

CEcolors(...)
}
\arguments{
\item{colors}{vector of numbers corresponding to EMI colors}

\item{accentLoc}{numbers specifying index to replace with accent color}

\item{ramp}{number of colors to interpolate - only works if length of colors
is 1 (fade to black) or 2 (interpolate). Default is for no ramping.}

\item{accent}{accent color. EMI Orange by default}

\item{order}{ordered vector of colors. Default is the EMI color set.}
}
\description{
Returns a set of colors from the EMI template. Can return a list of colors,
the EMI accent color (orange), or a faded range of colors.
}
\examples{
barplot(rep(1,4), col=EMIcolors(rep(1,4), accentLoc = 3))
barplot(rep(1,10), col=EMIcolors(1:10)) # all EMI colors
barplot(rep(1,5), col=EMIcolors(1:2, ramp=5)) # for satisfaction
}

