% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingFunctions.R
\name{EMIcolorsShow}
\alias{EMIcolorsShow}
\title{Show EMI Colors}
\usage{
EMIcolorsShow(colors = 1:10, ...)
}
\arguments{
\item{colors}{list of colors, by number, to show. Defaults to all 10 colors}
}
\description{
Displays a barplot of EMI's color scheme.
}

