% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingFunctions.R
\name{EMIknitSetup}
\alias{EMIknitSetup}
\title{EMI KnitR Setup}
\usage{
EMIknitSetup(tempPath = NULL)
}
\arguments{
\item{tempPath}{folder path to temporarily save files during compilation.
Currently only used for figures.}
}
\description{
Sets up formatting to knit (or compile) an R notebook. Optimized for HTML
outputs, although it should work for PDF or Word outputs as well.
}

