% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reportingFunctions.R
\name{EMIplot}
\alias{EMIplot}
\title{EMI Plot}
\usage{
EMIplot(...)
}
\arguments{
\item{...}{parameters for ggplot()}
}
\description{
Sets up a ggplot with the EMI formatting defaults - excluding colors
}
\details{
Typical EMI conventions for figures include:

- No title
- Size 9 font, Proxima Nova Regular
- Width of 6.5 in. (must be specified using ggsave(width = 6.5))
- Major gridlines only

Note: attempts to load and use Proxima Nova font. Defaults to Arial if
Proxima Nova cannot be loaded.
}

