% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingFunctions.R
\name{EMIsave}
\alias{EMIsave}
\title{EMI Save Figure}
\usage{
EMIsave(filename, ...)
}
\arguments{
\item{filename}{name of the file to save the plot}

\item{...}{other parameters passed to ggsave}
}
\description{
Saves the specified figure using ggsaev using EMI defaults. Defauts include
6in x 4in size and other ggsave defaults.
}

