% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importFunctions.R
\name{excel2csv}
\alias{excel2csv}
\title{Convert Excel File to .csv}
\usage{
excel2csv(infile, sheets = NULL, outfileNames = "number", ...)
}
\arguments{
\item{infile}{file path to input excel file (.xls, .xlsx, or .xlsm)}

\item{sheets}{number vector or character vector describing the sheets to save.
Default is all sheets are saved.}

\item{outfileNames}{output file string, or the string 'number' or 'name' to
specify the file names. Default is 'number', and output files are in
the format '<infile>_<sheetNumber>.csv'}
}
\description{
Saves an excel file (.xls, .xlsx, .xlsm) to csv files. Options specify whether to
save multiple sheets to separate csv files and what the output file names
should be.
}

