% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{expandLRpairs}
\alias{expandLRpairs}
\title{Generating cell type-specific ligand receptor pair network}
\usage{
expandLRpairs(lr.table, datlist, celltypes)
}
\arguments{
\item{lr.table}{List of all ligand receptor pairs and their pair information}

\item{datlist}{List of all cell type-specific gene expression matrices}

\item{celltypes}{List of all cell types in the dataset}
}
\value{
List including igraph ligand receptor network and data table of
ligand receptor pairs across all cell types
}
\description{
Network used for cabernet algorithm. Nodes represent ligand and receptor genes from
all cell types. Edges represent a literature-supported ligand receptor pair.
}
