% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{fileExtn}
\alias{fileExtn}
\title{### File System Functions
File Extension}
\usage{
fileExtn(fName)
}
\arguments{
\item{fName}{file name}
}
\description{
Returns file extenstion (ex. "csv" or "xls")
}

