% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_subfunctions.R
\name{findAdjacentCommEdges}
\alias{findAdjacentCommEdges}
\title{Clustering using label propagation}
\usage{
findAdjacentCommEdges(netlist, cgenes, comms)
}
\arguments{
\item{net}{LR network}

\item{clu}{Components in the network}

\item{clu.labeled}{Communities that contain at least two EC receptors}

\item{labelednodes}{List of EC receptors}
}
\value{
Communities identified using label propagation
}
\description{
This is the initial clustering step in cabernet. The network is
seeded with receptors that had a high eigenvector centrality
measurement to produce clusters that reflect downstream signaling
activity seen in the dataset.
}
