% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importFunctions.R
\name{fuzzyMerge}
\alias{fuzzyMerge}
\title{Fuzzy Matching for Merging Data Frames}
\usage{
fuzzyMerge(dfX, dfY, by = intersect(names(dfX), names(dfY))[1], byX = by,
  byY = by, costs = list(ins = 1, del = 0.25, sub = 1), distance = c(1, 2,
  3, 5, 7, 10, 15, 20), keepOriginal = FALSE, ...)
}
\arguments{
\item{dfX}{first data frame to match. The returned data frame will have the
same number of rows as this data frame.}

\item{dfY}{second data frame to match. Note: there should be no duplicates in
the matching column in this data frame!}

\item{by}{column name (or number) in data frames to use for matching. Can only
be one column! By default, it is the first matching column name in dfX and dfY}

\item{byX}{column name in dfX if column names are different}

\item{byY}{column name in dfY if column names are different}

\item{distance}{vector of maximum distances for fuzzy matching. See agrep for
details. Length corresponds to the number of matching iterations.}

\item{keepOriginal}{if True, adds the column "Original" in the final data frame
which contains vector.}

\item{...}{parameters sent to agrep for fuzzy matching}
}
\value{
a data frame with the same length as vector and the same columns as df.
The matched column will have the same name as col.
}
\description{
Merges two data frames using one shared column. Left merges only!
Direct matches are checked first, followed by multiple sets of
fuzzy matches. A random match is chosen if multiple values match.
}

