% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{getCSS}
\alias{getCSS}
\title{Get EMI CSS File}
\usage{
getCSS(dir)
}
\arguments{
\item{dir}{directory to copy the CSS file to. It should be the same as the
directory of the analysis file.}
}
\description{
Adds the default EMI CSS file to the chosen directory. Necessary for use in
generating "pretty" HTML files, for example using R Markdown (Rmd).
}

