% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingFunctions.R
\name{histFreqs}
\alias{histFreqs}
\title{Histogram Frequencies}
\usage{
histFreqs(v, increment, others = c(), ...)
}
\arguments{
\item{v}{vector to analyze}

\item{increment}{number to increment histogram bins (ex. 1)}

\item{others}{non-numeric levels to remove}
}
\description{
Turns a vector (or factor) into a data frame used for a table.
Input must be turned into numeric once 'others' are removed.
}

