% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{loadProximaNova}
\alias{loadProximaNova}
\title{Load Proxima Nova Font}
\usage{
loadProximaNova()
}
\description{
Imports the Proxima Nova font (found in the REMI directory on the secure
server) and loads it for pdf and postscript graphics. Returns TRUE on
success.
}

