% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_subfunctions.R
\name{logGaus}
\alias{logGaus}
\title{Calculating log Gaussian}
\usage{
logGaus(S, K, n)
}
\arguments{
\item{S}{Covariance matrix}

\item{K}{Inverse covariance matrix}

\item{n}{Sample size}
}
\value{
Log Gaussian calculation
}
\description{
Calculating log Gaussian
}
