% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_subfunctions.R
\name{makePPINetwork}
\alias{makePPINetwork}
\title{Making network function}
\usage{
makePPINetwork(data, filt.genes, pathgenes, c, lrmat, ppi = g.biogrid)
}
\arguments{
\item{data}{Gene expression data}

\item{filt.genes}{Genes that are nodes in the network}

\item{pathgenes}{List of KEGG pathway genes TO BE REMOVED}

\item{c}{Name of cell type to label the genes in the network}
}
\value{
A network
}
\description{
Given the dataset and genes, make a protein-protein interaction network
using BioGRID as a base network
}
