% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{myBasename}
\alias{myBasename}
\title{myBasename}
\usage{
myBasename(fName, includeFolders = FALSE, includeExtn = FALSE)
}
\arguments{
\item{fName}{full file path}

\item{includeFolders}{default is FALSE}

\item{includeExtn}{default is FALSE}
}
\description{
By default, returns the file name without the full path or the file extension.
It can include folders and include the file extension.
}

