% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generalFunctions.R
\name{mySplit}
\alias{mySplit}
\title{mySplit}
\usage{
mySplit(string, split, idx = NA, reverseIdx = FALSE, collapse = FALSE,
  ...)
}
\arguments{
\item{string}{input character vector to split}

\item{split}{the separator used to split string}

\item{reverseIdx}{if TRUE, reference indices from the end of the string.
Default is FALSE}

\item{collapse}{if TRUE and there are multiple indices,
collapses all indices into 1 string using split. Default is FALSE}

\item{...}{extra arguments passed to strsplit()}

\item{index}{optional, indices of split string to return.
Default returns the full split string}
}
\value{
character vector of split strings.
}
\description{
Splits a string by a separator. Similar to strsplit, but returns a vector
instead of a list. Works well for a single string, or for a vector if
you include the index that you want returned.
}
\examples{
mySplit('This_That', '_')
mySplit('This_That', '_', 2)
mySplit('This_That_Another', '_', -1, reverseIdx = TRUE)
}

