% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingFunctions.R
\name{myTable}
\alias{myTable}
\title{myTable}
\usage{
myTable(table, name = "X", sort = T, percent = F, total = percent,
  sum = NULL, remove = c(), bottom = c())
}
\arguments{
\item{table}{any 1-D table, or an object that can be converted to a two
column data frame}

\item{name}{string for first column header}

\item{sort}{sort table from highest to lowest (default is TRUE)}

\item{percent}{add column indicating percentages (default is FALSE)}

\item{total}{add row indicating total (default is FALSE)}

\item{sum}{total number to base percentages (default is the sum of all values)}

\item{remove}{vector of row names to remove (ex. NA)}

\item{bottom}{vector of row names to keep at the bottom of the table (ex. Other)}
}
\description{
Creates a 1-D table (i.e. from 1 factor) in a nice format. Useful for converting
to a printable table or for converting to a figure (using EMIbar).
}

