% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reportingFunctions.R
\name{myTableMulti}
\alias{myTableMulti}
\title{myTableMulti}
\usage{
myTableMulti(table, name = "", rowNames = NULL, colNames = NULL,
  sortRow = NULL, sortCol = NULL, percent = "", totalRow = F,
  totalCol = F, totalN = F, revRow = F, revCol = F, rmCol = c(),
  botCol = c(), rmRow = c(), botRow = c(), ...)
}
\arguments{
\item{table}{any 2-D table, or data frame that resembles a table}

\item{name}{Left-most column name (describing rows). Defaults to ""}

\item{rowNames}{names for table rows (pre-sorted) - defaults to table row names}

\item{colNames}{names for table columns (pre-sorted), EXCLUDING leftmost row}

\item{sortRow}{vector of column names/numbers to sort by (highest to lowest)}

\item{sortCol}{vector of column names/numbers to sort by (left to right)}

\item{percent}{convert all values to percentages.
Can evaluate percentages by 'row', 'col', or 'all'}

\item{totalRow}{add row indicating total (default is FALSE)}

\item{totalCol}{add column indicating total (default is FALSE)}

\item{totalN}{add (n=#) to row/col names instead of a new row/column (default is FALSE)}

\item{revRow}{reverse order of rows}

\item{revCol}{reverse order of columns}

\item{rmCol}{vector of column names to remove}

\item{botCol}{vector of column names to keep at the right of the table}

\item{rmRow}{vector of row names to remove}

\item{botRow}{vector of row names to keep at the bottom of the table}
}
\description{
Creates a 2-D table (i.e. from 2 factors) in a nice format. Useful for converting
to a printable table or determining percentages or totals.
}

