% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisFunctions.R
\name{outliers}
\alias{outliers}
\title{Find Outliers}
\usage{
outliers(v, deviations = 5, method = "normal")
}
\arguments{
\item{v}{input vector}

\item{deviations}{number of standard deviations to determine if a data point is an
outlier. Default is 5.}

\item{method}{method used to determine outliers. Default is to assume a
normal distribution}
}
\description{
Determines if there are any outliers in a set of data. By default, it assumes
a normal distribution and removes data beyond 5 std. deviations.
}

