% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingFunctions.R
\name{printList}
\alias{printList}
\title{Print Knitted List
Prints a vector of strings as a list in KnitR format. Works for
markdown and latex output formats.}
\usage{
printList(x, out.format = knitr::opts_knit$get("out.format"),
  environment = "itemize", marker = NULL)
}
\arguments{
\item{x}{vector of strings}

\item{out.format}{either 'markdown' or 'latex'.}

\item{environment}{used for latex formatting}

\item{marker}{used for latex formatting}
}
\description{
Print Knitted List
Prints a vector of strings as a list in KnitR format. Works for
markdown and latex output formats.
}

