% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingFunctions.R
\name{printTable}
\alias{printTable}
\title{Print Table}
\usage{
printTable(table, file = NULL, type = "HTML", include.rownames = FALSE,
  addCommas = FALSE, ...)
}
\arguments{
\item{table}{table-like R object (best results are with data frames)}

\item{file}{file path to save table. Default does not save the table}

\item{type}{format of output (default is HTML)}

\item{include.rownames}{Should row names be printed? (default is FALSE)}

\item{addCommas}{converts numbers to strings (1234 -> "1,234")}

\item{...}{extra parameters sent to xtable(table, ...)}
}
\description{
Prints and saves table or data frame with specified formats.
}

