% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysisFunctions.R
\name{pTest}
\alias{pTest}
\title{### Statistical Tests
pTest}
\usage{
pTest(v1, v2, type = "", out = "p", ...)
}
\arguments{
\item{v1}{numeric vector}

\item{v2}{numeric vector}

\item{type}{Specifies type of vector if v1 and v2 are not numeric.}

\item{out}{determines return value. Default ('p') returns the p statistic.}

\item{...}{Other parameters to t.test}
}
\description{
Performs a t-test (using t.test()) to compare two numeric vectors. Can also
test ordinal factors (ex. for satisfaction) using type='satis5'
}

