% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{remi}
\alias{remi}
\title{Cabernet algorithm}
\usage{
remi(cellmarkers, dat.list, seed = 30, numgenes = 2, cutoff = 1)
}
\arguments{
\item{dat.list}{List of gene expression for each cell type}

\item{seed}{Random seed number to ensure reproducible results}

\item{netlist}{List of LR network and table}

\item{labelednodes}{List of high eigenvector centrality receptors}
}
\value{
List of cabernet predictions
}
\description{
Identifies communities in LR network and then performs graphical lasso
on the communites in the network to predict significant ligand
receptor interactions in the dataset of interest
}
