% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REMI_C.R
\name{REMI_C}
\alias{REMI_C}
\title{REMI}
\usage{
REMI_C(stringname1, stringname2, nstep = 100, epsilon = 0.05,
  bandwidth = 200, shrinkagefactor = 0.9)
}
\arguments{
\item{stringname1}{File name w/ extension name for summary statistics with columns in the order of rs name, major allele, minor allele, estimated effect size, estimated standard error squared, sample size.}

\item{stringname2}{File name w/o extension name for reference panel data in the format of PLINK bfile.}

\item{nstep}{Number of steps for selecting tuning parameter. Default is 100.}

\item{epsilon}{Minimum value of tuning parameter in the ratio to the largest tuning parameter. Default is 0.05.}

\item{bandwidth}{The size of banded correlation to be calculated from reference panel. Default is 200.}

\item{shrinkagefactor}{The shrinkage amount to the estimated correlation matrix from reference panel. When shrinkagefactor is 1, there is no shrinkage. Default is 0.9.}
}
\value{
List of model parameters including major allele (A1), solution path (beta).
}
\description{
Fit REMI-C using marginal information.
}
\details{
\code{REMI_C} fits the REMI-C with marginal information. It requires to provide the reference panel data for SNPs in PLINK bfile format. The summary statistics in file \code{stringname1} and the reference panel is in file \code{stringname2}. This function is primarily used for comparison with REMI_R.
}