% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{REMIPlot}
\alias{REMIPlot}
\title{Making chord diagram for REMI results}
\usage{
REMIPlot(
  interactome,
  type = "chord",
  grid.col = NULL,
  size = 10,
  thres = 0,
  selectcell = NULL,
  legend = FALSE
)
}
\arguments{
\item{interactome}{Predicted interactome}

\item{type}{Plot type, either chord or alluvial}

\item{grid.col}{Optional: color option for cell type}

\item{size}{Size of the font}

\item{thres}{Remove LR pairs below this weight threshold}

\item{selectcell}{Particular cell to highlight in the chord diagram}

\item{legend}{Plot legend}
}
\value{
Chord Diagram highlighting proportion of cell-types in interactome
}
\description{
Making chord diagram for REMI results
}
