% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisFunctions.R
\name{sample.size}
\alias{sample.size}
\title{Sample Size Calculator}
\usage{
sample.size(N, P, A, cv)
}
\arguments{
\item{N}{Population Size}

\item{P}{Level of Confidence (~90 \%)}

\item{A}{Level of Precision (~10 \%)}

\item{cv}{Coefficient of Variation (~0.5)}
}
\description{
Determines the required sample size based on the population size, levels of
confidence and precision, and coefficient of variation (CV) of the variable of
interest.
}

