% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysisFunctions.R
\name{sdFromMean}
\alias{sdFromMean}
\title{Standard Deviations from Mean}
\usage{
sdFromMean(confidence)
}
\arguments{
\item{Confidence}{percentage (as a decimal)}
}
\description{
Determines the number of standards deviations from the mean, given a certain
confidence. For example, 0.95 (95% confidence) will return 1.96.
}

