% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{setupSingleCell}
\alias{setupSingleCell}
\title{Creating single cell REMI object}
\usage{
setupSingleCell(
  obj,
  sample.col,
  celltype.col,
  remove.markers = NULL,
  gene.select = NULL,
  assay = "integrated",
  filter = T,
  thres = 0,
  expthres = 0.1
)
}
\arguments{
\item{obj}{REMI object}

\item{celltype.col}{what cell types to measure}

\item{remove.markers}{cell types that should not be used for calculations}

\item{assay}{= Seurat assay}

\item{filter}{= remove low expressed ligand and receptors}

\item{thres}{= average assay expression filter}

\item{gene.selct}{genes of interest}
}
\value{
Single cell REMI object that can be used by the algorithm
}
\description{
Creating single cell REMI object
}
