% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R, R/load_functions_old.R
\name{SingleToBulk}
\alias{SingleToBulk}
\title{Averaging across expression levels}
\usage{
SingleToBulk(obj, assay, samplecol, celltypecol)

SingleToBulk(obj, assay, samplecol, celltypecol)
}
\arguments{
\item{obj}{REMI object}

\item{assay}{what Seurat assay to use for calculations}
}
\value{
Average expressiona cross patients

Average expressiona cross patients
}
\description{
Averaging across expression levels

Averaging across expression levels
}
