% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analysisFunctions.R
\name{weighted.mean}
\alias{weighted.mean}
\title{Weighted Mean}
\usage{
weighted.mean(x, w = NULL, xLoc = 2, wLoc = 3, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector. Alternatively, x can be a matrix-like object that
includes data for the vector and the weights.}

\item{w}{a numeric vector of weights.}

\item{xLoc}{if x is matrix-like, this specifies the column containing vector
information. Can be the column number or name.}

\item{wLoc}{if x is matrix-like, this specifies the column containing weight
information. Can be the column number or name.}
}
\description{
Returns a weighted mean given a vector and weights.
}
\details{
For 'tapply' or 'by' compatibility, the function can take a matrix or data frame
as long as the vector and weight column locations are specified.
}

