% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{cleaningOutput}
\alias{cleaningOutput}
\title{Cleaning Output}
\usage{
cleaningOutput(input, netlist)
}
\arguments{
\item{obj}{REMI object}
}
\value{
Cleaned up list of LR pairs with metadata displayed in columns
}
\description{
Identifies communities in LR network and then performs graphical lasso
on the communites in the network to predict significant ligand
receptor interactions in the dataset of interest
}
