% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_subfunctions.R
\name{EBIC}
\alias{EBIC}
\title{EBIC calculation}
\usage{
EBIC(S, K, n, E, gamma = 1, countDiagonal = FALSE)
}
\arguments{
\item{S}{Covariance matrix}

\item{K}{Inverse covariance matrix}

\item{n}{Sample number}
}
\value{
EBIC number
}
\description{
EBIC calculation
}
