% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/reportingFunctions.R
\name{EMIbar}
\alias{EMIbar}
\title{EMI Bar Graph 1-D}
\usage{
EMIbar(data, colors = EMIcolors(1), dataCols = 1:2, ...)
}
\arguments{
\item{data}{a data frame with columns X (bar names), and Count (bar size).
Works best with data from the myTable function.}

\item{colors}{vector of colors with length of 1 or the same length as data.}

\item{dataCols}{length 2 vector specifying the location of data within the
data frame. The first column should have category info, the second should have
counts.
Default colors are all EMI Blue (EMIcolors(1))}
}
\description{
Creates generic EMI bar graph. Takes a data frame (or data from myTable) to
create a 1-D plot. Colors are customizable.
}

