% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingFunctions.R
\name{EMIbarMulti}
\alias{EMIbarMulti}
\title{EMI Bar Graph 2-D}
\usage{
EMIbarMulti(data, colors = NULL)
}
\arguments{
\item{data}{a data frame with columns X (bar names), Fill (bar color),
and Count (bar size). Works best with data from the myTableMulti function.}

\item{colors}{vector of colors with length of 1 or the same length as the
number of Fill options. Default colors are all EMI Blue (EMIcolors(1))}

\item{...}{parameters passed to geom_bar}
}
\description{
Creates generic EMI bar graph. Takes a data frame (or data from myTableMulti) to
create a 2-D plot. Colors are customizable.
}
\examples{
df = data.frame(Factor1 = sample(LETTERS[1:4], 100, replace = T),
                Factor2 = sample(LETTERS[25:26], 100, replace = T),
                Numbers = sample(1:100))

# EMIbarMulti
t = bydf(df$Numbers, list(df$Factor1, df$Factor2), mean)
names(t) = c('X', 'Fill', 'Count')
EMIbarMulti(t)
}

