% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_subfunctions.R, R/load_subfunctions_old.R
\name{getUpperTri}
\alias{getUpperTri}
\title{Given graphical lasso matrix form output, extract edges by
looking at rows vs. column elements}
\usage{
getUpperTri(fit, round = TRUE)

getUpperTri(fit, round = TRUE)
}
\arguments{
\item{fit}{glasso matrix output}
}
\value{
List form of predicted edges

List form of predicted edges
}
\description{
Given graphical lasso matrix form output, extract edges by
looking at rows vs. column elements

Given graphical lasso matrix form output, extract edges by
looking at rows vs. column elements
}
