% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_subfunctions.R
\name{loglik_ave}
\alias{loglik_ave}
\title{Log Likelihood}
\usage{
loglik_ave(data, theta)
}
\arguments{
\item{data}{Community gene expression matrix}

\item{theta}{Parameter for log-likelihood calculation}
}
\value{
A network
}
\description{
Log Likelihood
}
