% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_subfunctions.R, R/load_subfunctions_old.R
\name{null_density}
\alias{null_density}
\title{Generate null density given covariance matrix}
\usage{
null_density(r, S, i, j, n)

null_density(r, S, i, j, n)
}
\arguments{
\item{r}{density}

\item{S}{covariance matrix}

\item{i}{index for ligand}

\item{j}{index for receptor}

\item{n}{sample size}
}
\value{
Null density

Null density
}
\description{
Generate null density given covariance matrix

Generate null density given covariance matrix
}
