% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{pickECgenes}
\alias{pickECgenes}
\title{Calculating eigenvector centrality for each receptor given downstream
protein protein-interaction signaling network.}
\usage{
pickECgenes(
  lrnet,
  dat.list,
  pgenelist,
  numgenes,
  cutoff,
  ppi,
  seed = 10,
  verbose = T
)
}
\arguments{
\item{lrnet}{List of cell type-specific LR pairs}

\item{dat.list}{List of gene expression for each cell type}

\item{pgenelist}{TO BE REMOVED}

\item{numgenes}{Number of genes user wants to seed clustering algorithm with}
}
\value{
List of receptors with high scoring eigenvector centrality measurements
}
\description{
Calculating eigenvector centrality for each receptor given downstream
protein protein-interaction signaling network.
}
