% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R, R/load_functions_old.R
\name{remifiedGlasso}
\alias{remifiedGlasso}
\title{Calculate graphical lasso on each community to sparsify the network}
\usage{
remifiedGlasso(netlist, communities, dat.list, seednum, lambda, scale = F)

remifiedGlasso(netlist, communities, dat.list, seednum, lambda, scale = F)
}
\arguments{
\item{netlist}{List of LR network and table}

\item{communities}{LR network communities}

\item{seednum}{Number of random seed to ensure reproducible results}

\item{datlist}{List of cell type-specific gene expression matrices}
}
\value{
List of edges remaining in LR network

List of edges remaining in LR network
}
\description{
Calculate graphical lasso on each community to sparsify the network

Calculate graphical lasso on each community to sparsify the network
}
